// client/src/data/itemSets.ts
export interface ItemSetDef {
  setId: number;
  setName: string;
  pieces: { id: number; name: string }[];
}